55 monster mix (#56)
* Refactor: Structure for considering Mix-monsters

#55

~ 몬스터끼리 먹이는 개념 감안하여 구조 재정립중

* Refactor: Structure for considering Mix-monsters

#55

~ 몬스터 구조 변경에 따른 관련 코드들 수정

* Refactor: Structure for considering Mix-monsters

#55

~ 몬스터 경험치화 대비 구조 재정립 완료
~ 관련된 코드 수정 완료 -> 작동 확인 완료

+++ SlimeGG/Assets/Scripts/Battles/Controller/FieldController.cs
@@ -44,7 +44,7 @@ public class FieldController : MonoBehaviour
         monster.SetParent(monsterContainer);
     }
 
-    public void setMonsterInPosition(Transform monster, int side, int[] setPos)
+    public void setMonsterInPosition(Transform monster, int side, float[] setPos)
     {
         setMonsterInPosition(monster);
         monster.localPosition = new Vector3((side == 1 ? 1 : -1) * (1 + (setPos[0])), setPos[1], 0f);
